fix 64-bit systems failures, bump version to 2.1.12
2005-08-30  Paolo Bonzini  <bonzini@gnu.org>

	* tests/geometry.st: Make more resilient to #basicHash vagaries.
	* tests/geometry.ok: Adapt.

libgst:
2005-08-30  Paolo Bonzini  <bonzini@gnu.org>

	* libgst/mpz.c: Fix 64-bit cleanliness problem.

git-archimport-id: bonzini@gnu.org--2004b/smalltalk--devo--2.1--patch-36

+++ libgst/mpz.c
@@ -1541,7 +1541,7 @@ _gst_mpz_from_oop(gst_mpz *mpz, OOP srcOOP)
 
   if (IS_INT(srcOOP))
     {
-      int i = TO_INT(srcOOP);
+      long i = TO_INT(srcOOP);
       gst_mpz_realloc (mpz, 1);
       mpz->size = i < 0 ? -1 : 1;
       mpz->d[0] = i < 0 ? -i : i;
